django-app/requirements.txt
Django
onnxruntime
Pillow
torch
torchvision